+++ CHANGELOG.md
@@ -4,8 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [Unreleased]
+## [1.5.0] - 2018-07-25
+### Changed
+- Using iOS SDK with ARKit
 
+### Added
+- AR Tracking listener
+- Model Gesture listener
 
 ## [1.4.0] - 2018-05-30
 ### Changed